Distinguish TeX True and TeX False by constructor.
This commit is about the representation of TeX snippets. There are two
kinds of TeX snippets:

1. Snippets resulting from {-" ... "-} comments in the code. These
snippets used to be represented as (TeX True "...") and are now
represented as (TeX "...").

2. Snippets resulting from "..." parts in %format directives. These
snippets used to be represented by (TeX False "...") and are now
represented as (Replacement "...").

This commit prepares for extending the representation of TeX to account
for languages with a different comment syntax. The representation of
Replacement is independent of the comment syntax, so it makes sense to
separate the two forms of snippets more.